Cleanup/TODOs
Former-commit-id: 779a237c7bc6244d5f1a8656253584e51c3f81cb

+++ Intersect.Server/Classes/Entities/Entity.cs
@@ -540,7 +540,7 @@ namespace Intersect.Server.Entities
             var time = 1000f / (float)(1 + Math.Log(Stat[(int)Stats.Speed].Value()));
             if (Blocking)
             {
-                time += time * (float)Options.BlockingSlow;
+                time += time * Options.BlockingSlow;
             }
             return Math.Min(1000f, time);
         }
